🌱 refactor: split oversized mock and card test files
- Split CardRuntime-coverage.test.tsx (1045 lines) into focused test files:
  - CardRuntime.registry.test.tsx: Registry functions (registerDataHook, etc.)
  - CardRuntime.states.test.tsx: Loading, error, empty, and hook-missing states
  - CardRuntime.visualizations.test.tsx: List/status and table visualizations
  - CardRuntime.features.test.tsx: Drill-downs, renderers, search, pagination, filters, sorting, header, and inference
  - CardRuntime.setup.ts: Shared mocks, helpers, and setup logic

- handlers.compliance.ts already refactored as barrel file
  - Re-exports from: frameworks, gov, security, erm modules

- Preserved all test assertions and handler logic
- Imports and mocks consolidated in setup file for reuse

Addresses: #20996 #21006
