+++ .npm-cache/_cacache/index-v5/2c/63/82205759d5f095b57bf9a97361174cf0b8ef29d9e7c0de9e77ae04cbd1b6
@@ -1,0 +1,2 @@
+
+4a338eef722c78855fae5aecad66b3dc68b780e0	{"key":"make-fetch-happen:request-cache:https://registry.npmjs.org/vitest","integrity":"sha512-lTVhLXxxjczPTJ7IdG9W7byuWNnxlWMS6mwh5PaDdM/ZplT/lxJQaahu8nwaQEEzQ3boNH7ucexrZqatTZVnUg==","time":1784042726716,"size":1261108,"metadata":{"time":1784042726625,"url":"https://registry.npmjs.org/vitest","reqHeaders":{"accept":"application/vnd.npm.install-v1+json; q=1.0, application/json; q=0.8, */*"},"resHeaders":{"cache-control":"public, max-age=300","content-encoding":"gzip","content-type":"application/vnd.npm.install-v1+json","date":"Tue, 14 Jul 2026 15:25:26 GMT","etag":"W/\"23104d5644ddae8e7ddcc4d160bdc979\"","last-modified":"Mon, 06 Jul 2026 06:54:11 GMT","vary":"accept-encoding, accept"},"options":{"compress":true}}}
