+++ files-to-check.txt
@@ -1,0 +1,141 @@
+pkg/agent/broadcast.go
+pkg/agent/config/config.go
+pkg/agent/config/testing.go
+pkg/agent/config_compat.go
+pkg/agent/csrf.go
+pkg/agent/error_sanitizer.go
+pkg/agent/federation/actions.go
+pkg/agent/federation/provider.go
+pkg/agent/federation/providers/capi.go
+pkg/agent/federation/providers/capi_actions.go
+pkg/agent/federation/providers/clusternet.go
+pkg/agent/federation/providers/clusternet_actions.go
+pkg/agent/federation/providers/karmada.go
+pkg/agent/federation/providers/karmada_actions.go
+pkg/agent/federation/providers/kubeadmiral.go
+pkg/agent/federation/providers/kubeadmiral_actions.go
+pkg/agent/federation/providers/liqo.go
+pkg/agent/federation/providers/liqo_actions.go
+pkg/agent/federation/providers/ocm.go
+pkg/agent/federation/providers/ocm_actions.go
+pkg/agent/federation/registry.go
+pkg/agent/federation/types.go
+pkg/agent/httputil/handler_context.go
+pkg/agent/httputil/json.go
+pkg/agent/init.go
+pkg/agent/kagent/kagent.go
+pkg/agent/kagent_compat.go
+pkg/agent/kube/client.go
+pkg/agent/kube/context.go
+pkg/agent/kube/discovery.go
+pkg/agent/kube/sanitize.go
+pkg/agent/kube/testing.go
+pkg/agent/kube/validation.go
+pkg/agent/kube/vcluster.go
+pkg/agent/middleware_body_limit.go
+pkg/agent/preflight.go
+pkg/agent/procgroup_compat.go
+pkg/agent/procutil/procgroup_unix.go
+pkg/agent/procutil/procgroup_windows.go
+pkg/agent/prometheus.go
+pkg/agent/prompt_sanitize.go
+pkg/agent/prompts/prompts.go
+pkg/agent/prompts/shell_unix.go
+pkg/agent/prompts/shell_windows.go
+pkg/agent/prompts/system_prompts.go
+pkg/agent/protocol/messages.go
+pkg/agent/provider.go
+pkg/agent/provider_prompts.go
+pkg/agent/providers/network.go
+pkg/agent/providers/provider_antigravity.go
+pkg/agent/providers/provider_bob.go
+pkg/agent/providers/provider_claude.go
+pkg/agent/providers/provider_claudecode.go
+pkg/agent/providers/provider_claudedesktop.go
+pkg/agent/providers/provider_cline.go
+pkg/agent/providers/provider_codex.go
+pkg/agent/providers/provider_continue.go
+pkg/agent/providers/provider_copilotcli.go
+pkg/agent/providers/provider_cursor.go
+pkg/agent/providers/provider_gemini.go
+pkg/agent/providers/provider_geminicli.go
+pkg/agent/providers/provider_goose.go
+pkg/agent/providers/provider_groq.go
+pkg/agent/providers/provider_helpers.go
+pkg/agent/providers/provider_jetbrains.go
+pkg/agent/providers/provider_kagenti.go
+pkg/agent/providers/provider_local_openai_compat.go
+pkg/agent/providers/provider_openai.go
+pkg/agent/providers/provider_openai_compat.go
+pkg/agent/providers/provider_openrouter.go
+pkg/agent/providers/provider_openwebui.go
+pkg/agent/providers/provider_raycast.go
+pkg/agent/providers/provider_vscode.go
+pkg/agent/providers/provider_windsurf.go
+pkg/agent/providers/provider_zed.go
+pkg/agent/providers_compat.go
+pkg/agent/registry.go
+pkg/agent/server.go
+pkg/agent/server_ai_chat.go
+pkg/agent/server_ai_chat_stream.go
+pkg/agent/server_ai_mixed.go
+pkg/agent/server_ai_mixed_security.go
+pkg/agent/server_ai_tokens.go
+pkg/agent/server_ai_websocket.go
+pkg/agent/server_argocd.go
+pkg/agent/server_auth.go
+pkg/agent/server_cilium.go
+pkg/agent/server_console_cr.go
+pkg/agent/server_events_stream.go
+pkg/agent/server_exec.go
+pkg/agent/server_federation.go
+pkg/agent/server_gitops.go
+pkg/agent/server_gpu_health.go
+pkg/agent/server_health.go
+pkg/agent/server_helm.go
+pkg/agent/server_http.go
+pkg/agent/server_http_kubeconfig.go
+pkg/agent/server_http_resources.go
+pkg/agent/server_http_resources_config.go
+pkg/agent/server_http_resources_rbac_storage.go
+pkg/agent/server_http_resources_stream.go
+pkg/agent/server_http_resources_workloads.go
+pkg/agent/server_http_workloads.go
+pkg/agent/server_jaeger.go
+pkg/agent/server_lifecycle.go
+pkg/agent/server_nvidia.go
+pkg/agent/server_ops_clusters.go
+pkg/agent/server_ops_health.go
+pkg/agent/server_ops_insights.go
+pkg/agent/server_ops_predictions.go
+pkg/agent/server_ops_settings.go
+pkg/agent/server_ops_validation.go
+pkg/agent/server_rbac.go
+pkg/agent/server_routes.go
+pkg/agent/shell.go
+pkg/agent/shell_unix.go
+pkg/agent/shell_windows.go
+pkg/agent/tokentracker/flock_unix.go
+pkg/agent/tokentracker/flock_windows.go
+pkg/agent/tokentracker/tracker.go
+pkg/agent/updater/build.go
+pkg/agent/updater/build_helpers.go
+pkg/agent/updater/checker.go
+pkg/agent/updater/download.go
+pkg/agent/updater/execute.go
+pkg/agent/updater/git.go
+pkg/agent/updater/helpers.go
+pkg/agent/updater/platform_unix.go
+pkg/agent/updater/platform_windows.go
+pkg/agent/updater/poller.go
+pkg/agent/websocket_deadline.go
+pkg/agent/workers/prediction_analysis.go
+pkg/agent/workers/prediction_metrics.go
+pkg/agent/workers/prediction_worker.go
+pkg/agent/workers/registry.go
+pkg/agent/workers/workers_device_tracker.go
+pkg/agent/workers/workers_insight.go
+pkg/agent/workers/workers_metrics_history.go
+pkg/agent/workers_compat.go
+pkg/agent/workers_dcgm_scraper.go
+pkg/agent/workers_fanout.go
