+++ templates-header.txt
@@ -1,0 +1,64 @@
+/**
+ * Template generators for widget code generation
+ * Contains render function generators for each card type
+ */
+
+import { WIDGET_CARDS, WIDGET_STATS, WIDGET_TEMPLATES } from './widgetRegistry'
+import { generateWidgetStyles, generateWidgetShell } from './styleConverter'
+import { resolveWidgetEndpoint, generateWidgetCommand } from './codeGenerator.utils'
+
+// Generate render function for specific card type
+export function generateCardRenderFunction(cardType: string, displayName?: string): string {
+  const card = WIDGET_CARDS[cardType]
+  const title = displayName || card?.displayName || cardType
+
+  // All render functions use ({ output }) pattern — Übersicht passes curl stdout as output string
+  const parseBlock = `
+  let data = null;
+  let error = null;
+  try {
+    const trimmed = (output || '').trim();
+    if (!trimmed) {
+      error = 'No response';
+    } else if (trimmed.startsWith('<!') || trimmed.startsWith('<html')) {
+      error = 'Endpoint not available';
+    } else if (trimmed.includes('"error"')) {
+      try {
+        const parsed = JSON.parse(trimmed);
+        error = parsed.error || 'Load failed';
+      } catch {
+        error = 'Load failed';
+      }
+    } else {
+      data = JSON.parse(trimmed);
+    }
+  } catch (e) {
+    error = 'Parse error';
+  }
+
+  const containerStyle = {
+    ...styles.card,
+    position: 'absolute',
+    top: widgetPosition.top + 'px',
+    left: widgetPosition.left + 'px',
+    pointerEvents: 'auto',
+  };`
+
+  // Wrap card content in the draggable container
+  const wrapOpen = `
+  return (
+    <div style={{ position: 'relative', width: '100%', height: '100%' }}>
+      <div className="widget-container" style={containerStyle}>
+        <div className="drag-handle" style={styles.dragHandle} onMouseDown={handleDragStart} title="Drag to move">
+          <span style={styles.dragIndicator}>⋮⋮</span>
+        </div>`
+
+  const wrapClose = `
+      </div>
+    </div>
+  );`
+
+  const issueButton = `
+        <div style={{marginTop: '8px'}}><button style={styles.issueBtn} onClick={() => openIssue(error)}>Report Issue</button></div>`
+
+  switch (cardType) {
