+++ web/src/hooks/__tests__/useCachedLLMd.hooks.test.ts
@@ -1001,157 +1001,4 @@ describe('useCachedLLMd', () => {
       consoleError.mockRestore()
     })
   })
-
-  // ========================================================================
-  // fetchLLMdModels (exported async function)
-  // ========================================================================
-
-  describe('fetchLLMdModels', () => {
-    it('fetches InferencePool resources and maps to LLMdModel', async () => {
-      const poolItems = [
-        {
-          metadata: { name: 'pool-1', namespace: 'llm-d' },
-          spec: { selector: { matchLabels: { 'llmd.org/model': 'llama-3-70b' } } },
-          status: {
-            parents: [
-              { conditions: [{ type: 'Accepted', status: 'True' }] },
-            ],
-          },
-        },
-      ]
-
-      mockKubectlProxy.exec.mockResolvedValue(mockExecJson(poolItems))
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models).toHaveLength(1)
-      expect(models[0].name).toBe('llama-3-70b')
-      expect(models[0].status).toBe('loaded')
-      expect(models[0].cluster).toBe('c1')
-      expect(models[0].namespace).toBe('llm-d')
-    })
-
-    it('uses pool name when no model label in selector', async () => {
-      const poolItems = [
-        {
-          metadata: { name: 'my-pool', namespace: 'inference' },
-          spec: { selector: { matchLabels: {} } },
-          status: {},
-        },
-      ]
-
-      mockKubectlProxy.exec.mockResolvedValue(mockExecJson(poolItems))
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models[0].name).toBe('my-pool')
-    })
-
-    it('sets status to "stopped" when no Accepted condition', async () => {
-      const poolItems = [
-        {
-          metadata: { name: 'pool-1', namespace: 'llm-d' },
-          spec: {},
-          status: {
-            parents: [
-              { conditions: [{ type: 'Accepted', status: 'False' }] },
-            ],
-          },
-        },
-      ]
-
-      mockKubectlProxy.exec.mockResolvedValue(mockExecJson(poolItems))
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models[0].status).toBe('stopped')
-    })
-
-    it('sets status to "stopped" when no parents at all', async () => {
-      const poolItems = [
-        {
-          metadata: { name: 'pool-1', namespace: 'llm-d' },
-          spec: {},
-        },
-      ]
-
-      mockKubectlProxy.exec.mockResolvedValue(mockExecJson(poolItems))
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models[0].status).toBe('stopped')
-    })
-
-    it('calls onProgress with accumulated results', async () => {
-      mockKubectlProxy.exec.mockResolvedValue(
-        mockExecJson([
-          {
-            metadata: { name: 'pool-1', namespace: 'llm-d' },
-            spec: {},
-            status: {},
-          },
-        ]),
-      )
-
-      const onProgress = vi.fn()
-      const { fetchLLMdModels } = await loadModule()
-      await fetchLLMdModels(['c1', 'c2'], onProgress)
-
-      expect(onProgress).toHaveBeenCalled()
-    })
-
-    it('returns empty on non-zero exit code', async () => {
-      mockKubectlProxy.exec.mockResolvedValue({ exitCode: 1, output: '' })
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models).toEqual([])
-    })
-
-    it('handles cluster errors gracefully', async () => {
-      mockKubectlProxy.exec.mockRejectedValue(new Error('timeout'))
-
-      const consoleError = vi.spyOn(console, 'error').mockImplementation(() => {})
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1'])
-
-      expect(models).toEqual([])
-      consoleError.mockRestore()
-    })
-
-    it('suppresses demo mode errors without logging', async () => {
-      mockKubectlProxy.exec.mockRejectedValue(new Error('demo mode active'))
-
-      const consoleError = vi.spyOn(console, 'error').mockImplementation(() => {})
-      const { fetchLLMdModels } = await loadModule()
-      await fetchLLMdModels(['c1'])
-
-      expect(consoleError).not.toHaveBeenCalled()
-      consoleError.mockRestore()
-    })
-
-    it('returns models from multiple clusters', async () => {
-      let callCount = 0
-      mockKubectlProxy.exec.mockImplementation(async () => {
-        callCount++
-        return mockExecJson([
-          {
-            metadata: { name: `pool-${callCount}`, namespace: 'llm-d' },
-            spec: {},
-            status: {},
-          },
-        ])
-      })
-
-      const { fetchLLMdModels } = await loadModule()
-      const models = await fetchLLMdModels(['c1', 'c2'])
-
-      expect(models.length).toBe(2)
-    })
-  })
 })
