+++ web/src/hooks/__tests__/useStackDiscovery.advanced.test.ts
@@ -1,0 +1,798 @@
+import { describe, it, expect, vi, beforeEach } from 'vitest'
+import { renderHook, act, waitFor } from '@testing-library/react'
+
+// ── Mocks ────────────────────────────────────────────────────────────────────
+
+// vi.hoisted keeps these declarations available inside vi.mock factories,
+// which are themselves hoisted to the top of the module by vitest.
+const { mockGetDemoMode, mockExec } = vi.hoisted(() => ({
+  mockGetDemoMode: vi.fn(() => false),
+  mockExec: vi.fn(),
+}))
+
+vi.mock('../useDemoMode', () => ({
+  useDemoMode: () => ({ isDemoMode: false, toggleDemoMode: vi.fn(), setDemoMode: vi.fn() }),
+  getDemoMode: mockGetDemoMode,
+}))
+
+vi.mock('../../lib/kubectlProxy', () => ({
+  kubectlProxy: { exec: (...args: unknown[]) => mockExec(...args) },
+}))
+
+import { useStackDiscovery, stackToServerMetrics } from '../useStackDiscovery'
+import type { LLMdStack } from '../useStackDiscovery'
+
+// ── Constants mirrored from source ───────────────────────────────────────────
+
+const CACHE_KEY = 'kubestellar-stack-cache'
+const REFRESH_INTERVAL_MS = 120000
+
+// ── Helpers ──────────────────────────────────────────────────────────────────
+
+/** Build a kubectl JSON response wrapping items */
+function k8sResponse(items: unknown[], exitCode = 0) {
+  return { output: JSON.stringify({ items }), exitCode }
+}
+
+/** Build an error/connection-refused kubectl response */
+function errorResponse(msg = 'connection refused', exitCode = 1) {
+  return { output: msg, exitCode }
+}
+
+/** Empty successful response (no items) */
+const EMPTY_RESPONSE = k8sResponse([])
+
+/** Namespace-list response (jsonpath format) */
+function nsResponse(namespaces: string[]) {
+  return { output: namespaces.join(' '), exitCode: 0 }
+}
+
+/** Build a minimal pod resource with llm-d labels */
+function makePod(
+  name: string,
+  namespace: string,
+  role: string,
+  phase = 'Running',
+  ready = true,
+  extraLabels: Record<string, string> = {},
+) {
+  return {
+    metadata: {
+      name,
+      namespace,
+      labels: {
+        'llm-d.ai/role': role,
+        'pod-template-hash': 'abc123',
+        ...extraLabels,
+      },
+    },
+    status: {
+      phase,
+      containerStatuses: [{ ready }],
+    },
+  }
+}
+
+/** Build a minimal deployment resource */
+function makeDeployment(
+  name: string,
+  namespace: string,
+  replicas = 1,
+  readyReplicas = 1,
+  labels: Record<string, string> = {},
+) {
+  return {
+    metadata: { name, namespace, labels: {} },
+    spec: {
+      replicas,
+      template: { metadata: { labels } },
+    },
+    status: { replicas, readyReplicas, availableReplicas: readyReplicas },
+  }
+}
+
+/** Build a minimal InferencePool resource */
+function makePool(name: string, namespace: string) {
+  return {
+    metadata: { name, namespace },
+    spec: { selector: { matchLabels: {} } },
+  }
+}
+
+/** Build a minimal HPA resource */
+function makeHPA(name: string, namespace: string, min = 1, max = 3) {
+  return {
+    metadata: { name, namespace },
+    spec: { minReplicas: min, maxReplicas: max },
+    status: { currentReplicas: min, desiredReplicas: min },
+  }
+}
+
+/** Build a minimal WVA resource */
+function makeWVA(name: string, namespace: string, min = 1, max = 5) {
+  return {
+    metadata: { name, namespace },
+    spec: { minReplicas: min, maxReplicas: max },
+    status: { currentReplicas: min, desiredReplicas: min },
+  }
+}
+
+/** Build a minimal service resource with EPP naming */
+function makeEPPService(name: string, namespace: string) {
+  return {
+    metadata: { name, namespace },
+    spec: { ports: [{ port: 9002 }] },
+  }
+}
+
+/** Build a minimal Gateway resource */
+function makeGateway(name: string, namespace: string, hasAddress = true) {
+  return {
+    metadata: { name, namespace },
+    spec: { gatewayClassName: 'istio' },
+    status: hasAddress ? { addresses: [{ value: '10.0.0.1' }] } : {},
+  }
+}
+
+/**
+ * Configure mockExec to handle the standard 7 Phase-1 parallel calls,
+ * followed by the namespace list call, and optional Phase-2 deployment calls.
+ */
+function setupMockExec(options: {
+  pods?: unknown[]
+  pools?: unknown[]
+  services?: unknown[]
+  gateways?: unknown[]
+  hpas?: unknown[]
+  wvas?: unknown[]
+  vpas?: unknown[]
+  namespaces?: string[]
+  deploymentsByNs?: Record<string, unknown[]>
+  clusterError?: boolean
+}) {
+  const {
+    pods = [],
+    pools = [],
+    services = [],
+    gateways = [],
+    hpas = [],
+    wvas = [],
+    vpas = [],
+    namespaces = [],
+    deploymentsByNs = {},
+    clusterError = false,
+  } = options
+
+  mockExec.mockImplementation((args: string[]) => {
+    if (clusterError) {
+      return Promise.resolve(errorResponse('Unable to connect'))
+    }
+
+    const cmd = args.join(' ')
+
+    if (cmd.includes('pods') && cmd.includes('llm-d.ai/role')) {
+      return Promise.resolve(k8sResponse(pods))
+    }
+    if (cmd.includes('inferencepools')) {
+      return Promise.resolve(k8sResponse(pools))
+    }
+    if (cmd.includes('services')) {
+      return Promise.resolve(k8sResponse(services))
+    }
+    if (cmd.includes('gateway') && !cmd.includes('kgateway')) {
+      return Promise.resolve(k8sResponse(gateways))
+    }
+    if (cmd.includes('hpa')) {
+      return Promise.resolve(k8sResponse(hpas))
+    }
+    if (cmd.includes('variantautoscalings')) {
+      return Promise.resolve(k8sResponse(wvas))
+    }
+    if (cmd.includes('vpa')) {
+      return Promise.resolve(k8sResponse(vpas))
+    }
+    if (cmd.includes('namespaces')) {
+      return Promise.resolve(nsResponse(namespaces))
+    }
+    if (cmd.includes('deployments') && cmd.includes('-n')) {
+      const nsFlag = args.indexOf('-n')
+      const ns = nsFlag >= 0 ? args[nsFlag + 1] : ''
+      const deps = deploymentsByNs[ns] || []
+      return Promise.resolve(k8sResponse(deps))
+    }
+
+    return Promise.resolve(EMPTY_RESPONSE)
+  })
+}
+
+/**
+ * Wait for React state to settle after async operations.
+ * Uses a small delay to allow microtasks, Promises, and React state updates to flush.
+ * In vitest 4, act() with async callbacks can hang when hooks use setInterval,
+ * so we use a plain timeout instead.
+ */
+async function flush() {
+  // Allow enough time for chained async operations:
+  // useEffect -> refetch -> Phase 1 (Promise.all) -> Phase 2 (namespace query -> deployment batches)
+  await new Promise(resolve => setTimeout(resolve, 200))
+}
+
+// ── Test Suite ────────────────────────────────────────────────────────────────
+
+describe('useStackDiscovery', () => {
+  beforeEach(() => {
+    vi.clearAllMocks()
+    localStorage.clear()
+    mockGetDemoMode.mockReturnValue(false)
+    mockExec.mockResolvedValue(EMPTY_RESPONSE)
+  })
+
+  it('skips namespaces already discovered in Phase 1 during Phase 2', async () => {
+    setupMockExec({
+      pods: [makePod('pod-0', 'llm-d-ns', 'both')],
+      namespaces: ['llm-d-ns', 'inference-new'],
+      deploymentsByNs: {
+        'inference-new': [
+          makeDeployment('granite-server', 'inference-new', 1, 1, { 'llmd.org/model': 'granite' }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(2)
+    // Phase 2 should NOT re-query llm-d-ns (already in Phase 1)
+    const depCalls = mockExec.mock.calls.filter(
+      (c: unknown[]) => (c[0] as string[]).includes('deployments'),
+    )
+    const nsQueried = depCalls.map((c: unknown[]) => {
+      const args = c[0] as string[]
+      return args[args.indexOf('-n') + 1]
+    })
+    expect(nsQueried).not.toContain('llm-d-ns')
+    expect(nsQueried).toContain('inference-new')
+    unmount()
+  })
+
+  // ── 10. Deployment classification (EPP, prefill, decode, both) ─────────────
+
+  it('classifies deployment as EPP when name contains -epp', async () => {
+    setupMockExec({
+      pods: [],
+      namespaces: ['serving-ns'],
+      deploymentsByNs: {
+        'serving-ns': [
+          makeDeployment('model-epp', 'serving-ns', 1, 1),
+          makeDeployment('vllm-model', 'serving-ns', 2, 2, { 'app.kubernetes.io/name': 'vllm' }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    expect(result.current.stacks[0].components.epp).not.toBeNull()
+    expect(result.current.stacks[0].components.epp!.name).toBe('model-epp')
+    unmount()
+  })
+
+  it('classifies deployments with prefill/decode in the name', async () => {
+    setupMockExec({
+      pods: [],
+      namespaces: ['llm-d-pd'],
+      deploymentsByNs: {
+        'llm-d-pd': [
+          makeDeployment('granite-prefill', 'llm-d-pd', 3, 3, { 'llmd.org/model': 'granite-3b' }),
+          makeDeployment('granite-decode', 'llm-d-pd', 2, 2, { 'llmd.org/model': 'granite-3b' }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    const stack = result.current.stacks[0]
+    expect(stack.components.prefill.length).toBe(1)
+    expect(stack.components.decode.length).toBe(1)
+    expect(stack.hasDisaggregation).toBe(true)
+    expect(stack.model).toBe('granite-3b')
+    unmount()
+  })
+
+  it('role label takes precedence over deployment name (fix #13716)', async () => {
+    // Deployment named 'prefill-server' but explicitly labelled role=decode.
+    // Before the fix this landed in prefill because depName.includes('prefill')
+    // fired before role === 'decode' was checked.
+    setupMockExec({
+      pods: [],
+      namespaces: ['llm-d-pd'],
+      deploymentsByNs: {
+        'llm-d-pd': [
+          makeDeployment('prefill-server', 'llm-d-pd', 2, 2, { 'llm-d.ai/role': 'decode' }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    const stack = result.current.stacks[0]
+    expect(stack.components.decode.length).toBe(1)
+    expect(stack.components.prefill.length).toBe(0)
+    unmount()
+  })
+
+  // ── 11. Stack status computation ───────────────────────────────────────────
+
+  it('computes status=healthy when all components are running', async () => {
+    setupMockExec({
+      pods: [
+        makePod('prefill-0', 'ns1', 'prefill', 'Running', true),
+        makePod('decode-0', 'ns1', 'decode', 'Running', true),
+      ],
+      services: [makeEPPService('ns1-epp', 'ns1')],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    expect(result.current.stacks[0].status).toBe('healthy')
+    unmount()
+  })
+
+  it('computes status=unhealthy when no components are running', async () => {
+    setupMockExec({
+      pods: [
+        makePod('prefill-0', 'ns1', 'prefill', 'Pending', false),
+        makePod('decode-0', 'ns1', 'decode', 'Pending', false),
+      ],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    expect(result.current.stacks[0].status).toBe('unhealthy')
+    unmount()
+  })
+
+  // ── 12. Refresh interval ───────────────────────────────────────────────────
+
+  it('triggers silent refetch after REFRESH_INTERVAL_MS', async () => {
+    vi.useFakeTimers({ shouldAdvanceTime: true })
+    setupMockExec({
+      pods: [makePod('pod-0', 'ns1', 'both')],
+      namespaces: [],
+    })
+
+    const { unmount } = renderHook(() => useStackDiscovery(['c1']))
+    // Wait for initial fetch to complete
+    await vi.advanceTimersByTimeAsync(500)
+
+    const initialCallCount = mockExec.mock.calls.length
+
+    // Advance past the refresh interval to trigger a silent refetch
+    await vi.advanceTimersByTimeAsync(REFRESH_INTERVAL_MS + 500)
+
+    expect(mockExec.mock.calls.length).toBeGreaterThan(initialCallCount)
+    unmount()
+    vi.useRealTimers()
+  })
+
+  it('clears interval on unmount to prevent worker hangs', async () => {
+    vi.useFakeTimers({ shouldAdvanceTime: true })
+    setupMockExec({
+      pods: [makePod('pod-0', 'ns1', 'both')],
+      namespaces: [],
+    })
+
+    const { unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    unmount()
+
+    const callCountAfterUnmount = mockExec.mock.calls.length
+
+    await act(async () => {
+      await vi.advanceTimersByTimeAsync(REFRESH_INTERVAL_MS * 2)
+    })
+
+    // Allow at most 1 extra call from an in-flight async callback at unmount time
+    expect(mockExec.mock.calls.length).toBeLessThanOrEqual(callCountAfterUnmount + 1)
+    vi.useRealTimers()
+  })
+
+  // ── 13. Multiple clusters ─────────────────────────────────────────────────
+
+  it('processes multiple clusters sequentially and merges results', async () => {
+    mockExec.mockImplementation((args: string[], opts?: { context?: string }) => {
+      const cmd = args.join(' ')
+      const ctx = opts?.context || ''
+
+      if (cmd.includes('pods') && cmd.includes('llm-d.ai/role')) {
+        if (ctx === 'cluster-a') return Promise.resolve(k8sResponse([makePod('pa-0', 'ns-a', 'both')]))
+        if (ctx === 'cluster-b') return Promise.resolve(k8sResponse([makePod('pb-0', 'ns-b', 'both')]))
+        return Promise.resolve(EMPTY_RESPONSE)
+      }
+      if (cmd.includes('namespaces')) return Promise.resolve(nsResponse([]))
+      return Promise.resolve(EMPTY_RESPONSE)
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['cluster-a', 'cluster-b']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(2)
+    const ids = result.current.stacks.map(s => s.id)
+    expect(ids).toContain('ns-a@cluster-a')
+    expect(ids).toContain('ns-b@cluster-b')
+    unmount()
+  })
+
+  // ── 14. Cached merge (stale-while-revalidate) ─────────────────────────────
+
+  it('preserves cached component details when fresh fetch loses them', async () => {
+    const cachedStack: LLMdStack = {
+      id: 'merge-ns@c1',
+      name: 'merge-ns',
+      namespace: 'merge-ns',
+      cluster: 'c1',
+      components: {
+        prefill: [{
+          name: 'cached-prefill', namespace: 'merge-ns', cluster: 'c1',
+          type: 'prefill', status: 'running', replicas: 2, readyReplicas: 2,
+        }],
+        decode: [{
+          name: 'cached-decode', namespace: 'merge-ns', cluster: 'c1',
+          type: 'decode', status: 'running', replicas: 3, readyReplicas: 3,
+        }],
+        both: [],
+        epp: {
+          name: 'cached-epp', namespace: 'merge-ns', cluster: 'c1',
+          type: 'epp', status: 'running', replicas: 1, readyReplicas: 1,
+        },
+        gateway: null,
+      },
+      status: 'healthy',
+      hasDisaggregation: true,
+      model: 'granite-3b',
+      totalReplicas: 5,
+      readyReplicas: 5,
+      autoscaler: { type: 'HPA', name: 'my-hpa', minReplicas: 1, maxReplicas: 10 },
+    }
+
+    localStorage.setItem(CACHE_KEY, JSON.stringify({
+      stacks: [cachedStack],
+      timestamp: Date.now(),
+    }))
+
+    // Fresh fetch returns the namespace but pods API fails — components will be empty
+    setupMockExec({
+      pods: [],
+      pools: [makePool('merge-pool', 'merge-ns')],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    const stack = result.current.stacks.find(s => s.id === 'merge-ns@c1')!
+    expect(stack).toBeDefined()
+    expect(stack.components.prefill.length).toBe(1)
+    expect(stack.components.decode.length).toBe(1)
+    expect(stack.components.epp).not.toBeNull()
+    expect(stack.autoscaler?.type).toBe('HPA')
+    expect(stack.model).toBe('granite-3b')
+    unmount()
+  })
+
+  // ── 15. Pod status mapping ─────────────────────────────────────────────────
+
+  it('maps pod phase and container readiness to component status', async () => {
+    setupMockExec({
+      pods: [
+        makePod('running-pod', 'ns1', 'both', 'Running', true),
+        makePod('error-pod', 'ns1', 'both', 'Failed', false, { 'pod-template-hash': 'err-hash' }),
+      ],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    const components = result.current.stacks[0].components.both
+    const runningComp = components.find(c => c.readyReplicas > 0)
+    const errorComp = components.find(c => c.readyReplicas === 0)
+
+    expect(runningComp?.status).toBe('running')
+    expect(errorComp?.status).toBe('error')
+    unmount()
+  })
+
+  // ── 16. Unmount during active fetch ────────────────────────────────────────
+
+  it('does not crash when unmounted during an active fetch', async () => {
+    let resolveExec: ((v: unknown) => void) | null = null
+    mockExec.mockImplementation(() => new Promise(resolve => { resolveExec = resolve }))
+
+    const { unmount } = renderHook(() => useStackDiscovery(['c1']))
+
+    await flush()
+
+    unmount()
+
+    // Resolve the pending exec after unmount — should not throw
+    if (resolveExec) {
+      resolveExec(EMPTY_RESPONSE)
+    }
+  })
+
+  // ── 17. refetch function exposure ──────────────────────────────────────────
+
+  it('exposes a refetch function that triggers a non-silent refetch', async () => {
+    setupMockExec({
+      pods: [makePod('pod-0', 'ns1', 'both')],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    const callsBefore = mockExec.mock.calls.length
+
+    act(() => { result.current.refetch() })
+    await flush()
+
+    expect(mockExec.mock.calls.length).toBeGreaterThan(callsBefore)
+    unmount()
+  })
+
+  // ── 18. lastRefresh tracking ───────────────────────────────────────────────
+
+  it('updates lastRefresh after successful discovery', async () => {
+    setupMockExec({
+      pods: [makePod('pod-0', 'ns1', 'both')],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+
+    expect(result.current.lastRefresh).toBeNull()
+
+    await flush()
+
+    expect(result.current.lastRefresh).not.toBeNull()
+    expect(result.current.lastRefresh).toBeInstanceOf(Date)
+    unmount()
+  })
+
+  // ── 19. Namespace heuristic filtering ──────────────────────────────────────
+
+  it('filters Phase 2 namespaces using llm-d heuristics', async () => {
+    setupMockExec({
+      pods: [],
+      namespaces: [
+        'default',          // NOT an llm-d namespace
+        'kube-system',      // NOT an llm-d namespace
+        'vllm-production',  // IS (contains "vllm")
+        'inference-v2',     // IS (contains "inference")
+        'my-app',           // NOT
+      ],
+      deploymentsByNs: {
+        'vllm-production': [
+          makeDeployment('vllm-server', 'vllm-production', 1, 1, { 'app.kubernetes.io/name': 'vllm' }),
+        ],
+        'inference-v2': [
+          makeDeployment('llama-serving', 'inference-v2', 1, 1, { 'llmd.org/model': 'llama-2' }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(2)
+    const namespaces = result.current.stacks.map(s => s.namespace)
+    expect(namespaces).toContain('vllm-production')
+    expect(namespaces).toContain('inference-v2')
+    expect(namespaces).not.toContain('default')
+    expect(namespaces).not.toContain('kube-system')
+    unmount()
+  })
+
+  // ── 20. stackToServerMetrics ───────────────────────────────────────────────
+
+  it('converts a stack to LLMdServer[] with correct component types', () => {
+    const stack: LLMdStack = {
+      id: 'test-ns@c1',
+      name: 'test-ns',
+      namespace: 'test-ns',
+      cluster: 'c1',
+      components: {
+        prefill: [{
+          name: 'pf-0', namespace: 'test-ns', cluster: 'c1',
+          type: 'prefill', status: 'running', replicas: 2, readyReplicas: 2, model: 'granite',
+        }],
+        decode: [{
+          name: 'dc-0', namespace: 'test-ns', cluster: 'c1',
+          type: 'decode', status: 'running', replicas: 3, readyReplicas: 3, model: 'granite',
+        }],
+        both: [{
+          name: 'uni-0', namespace: 'test-ns', cluster: 'c1',
+          type: 'both', status: 'running', replicas: 1, readyReplicas: 1,
+        }],
+        epp: {
+          name: 'epp-0', namespace: 'test-ns', cluster: 'c1',
+          type: 'epp', status: 'running', replicas: 1, readyReplicas: 1,
+        },
+        gateway: {
+          name: 'gw-0', namespace: 'test-ns', cluster: 'c1',
+          type: 'gateway', status: 'running', replicas: 1, readyReplicas: 1,
+        },
+      },
+      status: 'healthy',
+      hasDisaggregation: true,
+      model: 'granite',
+      totalReplicas: 6,
+      readyReplicas: 6,
+    }
+
+    const servers = stackToServerMetrics(stack)
+
+    expect(servers.length).toBe(5)
+    expect(servers.filter(s => s.componentType === 'model').length).toBe(3)
+    expect(servers.filter(s => s.componentType === 'epp').length).toBe(1)
+    expect(servers.filter(s => s.componentType === 'gateway').length).toBe(1)
+
+    const eppServer = servers.find(s => s.componentType === 'epp')!
+    expect(eppServer.name).toBe('EPP Scheduler')
+
+    const gwServer = servers.find(s => s.componentType === 'gateway')!
+    expect(gwServer.name).toBe('Istio Gateway')
+    expect(gwServer.gatewayType).toBe('istio')
+  })
+
+  it('stackToServerMetrics uses stack model as fallback when component has no model', () => {
+    const stack: LLMdStack = {
+      id: 'fb-ns@c1',
+      name: 'fb-ns',
+      namespace: 'fb-ns',
+      cluster: 'c1',
+      components: {
+        prefill: [],
+        decode: [],
+        both: [{
+          name: 'server-0', namespace: 'fb-ns', cluster: 'c1',
+          type: 'both', status: 'running', replicas: 1, readyReplicas: 1,
+        }],
+        epp: null,
+        gateway: null,
+      },
+      status: 'healthy',
+      hasDisaggregation: false,
+      model: 'fallback-model',
+      totalReplicas: 1,
+      readyReplicas: 1,
+    }
+
+    const servers = stackToServerMetrics(stack)
+    expect(servers[0].model).toBe('fallback-model')
+  })
+
+  // ── 21. Stacks sorted: healthy first, then alphabetical ────────────────────
+
+  it('sorts stacks with healthy first, then by name', async () => {
+    mockExec.mockImplementation((args: string[]) => {
+      const cmd = args.join(' ')
+      if (cmd.includes('pods') && cmd.includes('llm-d.ai/role')) {
+        return Promise.resolve(k8sResponse([
+          makePod('pod-z', 'z-ns', 'both', 'Pending', false),
+          makePod('pod-a', 'a-ns', 'both', 'Running', true),
+        ]))
+      }
+      if (cmd.includes('namespaces')) return Promise.resolve(nsResponse([]))
+      return Promise.resolve(EMPTY_RESPONSE)
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(2)
+    // a-ns is healthy (running), z-ns is unhealthy — healthy comes first
+    expect(result.current.stacks[0].namespace).toBe('a-ns')
+    expect(result.current.stacks[1].namespace).toBe('z-ns')
+    unmount()
+  })
+
+  // ── 22. Pod role variants ──────────────────────────────────────────────────
+
+  it('recognizes prefill-server, decode-server, and vllm roles', async () => {
+    setupMockExec({
+      pods: [
+        makePod('ps-0', 'ns1', 'prefill-server'),
+        makePod('ds-0', 'ns1', 'decode-server', 'Running', true, { 'pod-template-hash': 'ds' }),
+        makePod('vl-0', 'ns1', 'vllm', 'Running', true, { 'pod-template-hash': 'vl' }),
+      ],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    const stack = result.current.stacks[0]
+    expect(stack.components.prefill.length).toBe(1)
+    expect(stack.components.decode.length).toBe(1)
+    expect(stack.components.both.length).toBe(1)
+    unmount()
+  })
+
+  // ── 23. VPA detection ──────────────────────────────────────────────────────
+
+  it('detects VPA as autoscaler when no WVA or HPA exist', async () => {
+    setupMockExec({
+      pods: [makePod('pod-0', 'vpa-ns', 'both')],
+      vpas: [{ metadata: { name: 'my-vpa', namespace: 'vpa-ns' } }],
+      namespaces: [],
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    expect(result.current.stacks[0].autoscaler?.type).toBe('VPA')
+    expect(result.current.stacks[0].autoscaler?.name).toBe('my-vpa')
+    unmount()
+  })
+
+  // ── 24. Deployment status mapping ──────────────────────────────────────────
+
+  it('maps deployment replicas/readyReplicas to correct component status', async () => {
+    setupMockExec({
+      pods: [],
+      namespaces: ['llm-d-status'],
+      deploymentsByNs: {
+        'llm-d-status': [
+          makeDeployment('healthy-model', 'llm-d-status', 3, 3, { 'app.kubernetes.io/name': 'vllm' }),
+          makeDeployment('degraded-model', 'llm-d-status', 3, 1, {
+            'app.kubernetes.io/name': 'vllm',
+            'pod-template-hash': 'deg',
+          }),
+        ],
+      },
+    })
+
+    const { result, unmount } = renderHook(() => useStackDiscovery(['c1']))
+    await flush()
+
+    expect(result.current.stacks.length).toBe(1)
+    const comps = result.current.stacks[0].components.both
+    const healthy = comps.find(c => c.name === 'healthy-model')
+    const degraded = comps.find(c => c.name === 'degraded-model')
+
+    expect(healthy?.status).toBe('running')
+    expect(degraded?.status).toBe('running') // readyReplicas > 0 => 'running'
+    unmount()
+  })
+
+  // ── 25. Return shape contract ──────────────────────────────────────────────
+
+  it('always returns the expected shape regardless of input', () => {
+    const { result, unmount } = renderHook(() => useStackDiscovery([]))
+
+    expect(result.current).toHaveProperty('stacks')
+    expect(result.current).toHaveProperty('isLoading')
+    expect(result.current).toHaveProperty('error')
+    expect(result.current).toHaveProperty('refetch')
+    expect(result.current).toHaveProperty('lastRefresh')
+    expect(Array.isArray(result.current.stacks)).toBe(true)
+    expect(typeof result.current.refetch).toBe('function')
+    unmount()
+  })
+})
